Move device_tree.rs
